integration test for configuration permissions for groups
