+++ dojo/templates/base.html
@@ -326,7 +326,7 @@
                                         <li><a href="{% url 'users' %}"> Users </a></li>
                                         {% endif %}
                                         {% if "auth.view_group"|has_configuration_permission:"staff" %}
-                                        <li><a href="{% url 'groups' %}"> Groups </a></li>
+                                        <li><a href="{% url 'groups' %}"  id="id_group_menu"> Groups </a></li>
                                         {% endif %}
                                     </ul>
                                 </li>
